Separate rules exception and relocate events header
