+++ production/inc/gaia/common.hpp
@@ -18,13 +18,13 @@
 namespace gaia
 {
 /**
- * \addtogroup Gaia
+ * \addtogroup gaia
  * @{
  */
 namespace common
 {
 /**
- * \addtogroup Common
+ * \addtogroup common
  * @{
  */
