+++ production/inc/gaia/direct_access/dac_array.hpp
@@ -16,7 +16,6 @@
 
 namespace gaia
 {
-
 namespace direct_access
 {
