+++ production/inc/gaia/direct_access/dac_iterators.hpp
@@ -15,17 +15,14 @@
 
 namespace gaia
 {
-
 /**
- * \addtogroup Gaia
+ * \addtogroup gaia
  * @{
  */
-
 namespace direct_access
 {
-
 /**
- * \addtogroup Direct
+ * \addtogroup direct_access
  * @{
  *
  * Implementation of Direct Access Classes. This provides a direct access API
