+++ production/inc/gaia/exception.hpp
@@ -14,13 +14,13 @@
 namespace gaia
 {
 /**
- * \addtogroup Gaia
+ * \addtogroup gaia
  * @{
  */
 namespace common
 {
 /**
- * \addtogroup Common
+ * \addtogroup common
  * @{
  */
