+++ production/inc/gaia/int_type.hpp
@@ -14,13 +14,13 @@
 namespace gaia
 {
 /**
- * \addtogroup Gaia
+ * \addtogroup gaia
  * @{
  */
 namespace common
 {
 /**
- * \addtogroup Common
+ * \addtogroup common
  * @{
  */
