+++ production/rules/event_manager/tests/test_component_init.cpp
@@ -9,7 +9,7 @@
 #include <gtest/gtest.h>
 
 #include "gaia/exceptions.hpp"
-#include "gaia/rules/rules.hpp"
+#include "gaia/rules/exceptions.hpp"
 
 #include "gaia_internal/db/db_test_base.hpp"
 #include "gaia_internal/rules/rules_config.hpp"
