+++ production/rules/event_manager/tests/test_event_manager.cpp
@@ -11,7 +11,7 @@
 #include <gtest/gtest.h>
 
 #include "gaia/direct_access/dac_base.hpp"
-#include "gaia/rules/rules.hpp"
+#include "gaia/rules/exceptions.hpp"
 
 #include "gaia_internal/db/db_catalog_test_base.hpp"
 #include "gaia_internal/db/triggers.hpp"
